Move the extension to callouts.py
